ref: drawState 이름 수정, canvasState 파일 분리
