+++ src/viewModel/CanvasViewModel.ts
@@ -2,9 +2,11 @@ import { CanvasModel } from "../model/CanvasModel";
 import React from "react";
 import { Observable } from "../core/Observable";
 import { Shape } from "../entity/Shape";
-import { DrawingState, ICanvasState, ResizeState } from "./CanvasState";
 import { CanvasEvent } from "./CanvasEvents";
 import { ZOrderAction } from "../model/CanvasModel";
+import { ICanvasState } from "./canvasState/CanvasState";
+import { DrawState } from "./canvasState/DrawState";
+import { ResizeState } from "./canvasState/ResizeState";
 
 export class CanvasViewModel extends Observable<any> {
   private model: CanvasModel;
@@ -15,7 +17,7 @@ export class CanvasViewModel extends Observable<any> {
   constructor(model: CanvasModel) {
     super();
     this.model = model;
-    this.state = new DrawingState(this); //default: 그리기 모드
+    this.state = new DrawState(this); //default: 그리기 모드
   }
 
   setState(state: ICanvasState) {
@@ -45,7 +47,7 @@ export class CanvasViewModel extends Observable<any> {
     this.clearShapes();
     this.clearSelectedShapes();
     this.setShapeType("rectangle"); // default 값으로
-    this.setState(new DrawingState(this));
+    this.setState(new DrawState(this));
     this.notifyShapesUpdated();
   }
 
@@ -147,7 +149,7 @@ export class CanvasViewModel extends Observable<any> {
         data: {
           currentState: this.state.constructor.name,
           drawingShape:
-            this.state instanceof DrawingState ? this.shapeType : undefined,
+            this.state instanceof DrawState ? this.shapeType : undefined,
         },
       };
     this.notify(event);
