+++ src/viewModel/canvasState/CanvasState.ts
@@ -1,0 +1,9 @@
+import React from "react";
+import { Shape } from "../../entity/Shape";
+
+export interface ICanvasState {
+  handleMouseDown(event: React.MouseEvent): void;
+  handleMouseMove(event: React.MouseEvent): void;
+  handleMouseUp(): void;
+  getCurrentShapes(): Shape[];
+}
